backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Cors
Werkzeug
SQLAlchemy
pymysql
python-dotenv
pandas